requirements.txt
Django>=4.2
 djangorestframework
 django-filter
drf-spectacular
drf-spectacular-sidecar
 drf-spectacular
